+++ src/spatRasterMultiple.h
@@ -35,7 +35,7 @@ class SpatRasterCollection {
 //		SpatExtent extent;
 		std::vector<std::string> names;
 		SpatRasterCollection() {};
-		SpatRasterCollection(std::string fname, std::vector<int> ids, bool useids);
+		SpatRasterCollection(std::string fname, std::vector<int> ids, bool useids, std::vector<std::string> options);
 //		void setExtent();
 		SpatExtent getExtent();
 		
@@ -78,7 +78,7 @@ class SpatRasterStack {
 		std::vector<std::string> long_names;
 		std::vector<std::string> units;
 		SpatRasterStack() {};
-		SpatRasterStack(std::string fname, std::vector<int> ids, bool useids);
+		SpatRasterStack(std::string fname, std::vector<int> ids, bool useids, std::vector<std::string> options);
 		SpatRasterStack(SpatRaster r, std::string name, std::string longname, std::string unit, bool warn=false);
 		std::vector<std::vector<std::vector<double>>> extractXY(std::vector<double> &x, std::vector<double> &y, std::string method);
 		std::vector<std::vector<std::vector<double>>> extractCell(std::vector<double> &cell);