+++ NEWS.md
@@ -1,4 +1,4 @@
-# version 1.7-60
+# version 1.7-61
 
 ## bug fixes
 
@@ -7,6 +7,7 @@ https://stackoverflow.com/questions/77304534/rspatial-terraflip-error-when-flipp
 - SpatRaster now correctly handles `NA & FALSE` and `NA | TRUE` [#1316](https://github.com/rspatial/terra/issues/1316) by John Baums
 - `set.names` wasn't working properly for SpatRasterDataset or SpatRasterCollection [#1333](https://github.com/rspatial/terra/pull/1333) by Derek Friend
 - `extract` with argument "layer" not NULL shifted the layers [#1332](https://github.com/rspatial/terra/issues/1332) by Ewan Wakefield
+- `terraOptions` did not capture "memmin" on [stackoverflow](https://stackoverflow.com/questions/77552234/controlling-chunk-size-in-terra) by dww
 
 
 ## enhancements
@@ -20,6 +21,8 @@ https://stackoverflow.com/questions/77304534/rspatial-terraflip-error-when-flipp
 
 - `k_means<SpatRaster>` method [#1314](https://github.com/rspatial/terra/issues/1314) by Agustin Lobo
 - `has.time<SpatRaster>` method 
+- new argument "raw=FALSE" to `rast`, `sds`, and `sprc` to ignore scale and offset [1354](https://github.com/rspatial/terra/issues/1354) by Insang Song
+
 
 # version 1.7-55